Draw bounding edges with horiz tangent lines
Before, I has points that would go to "infinity" to simulate tangent
lines. This has been changed to lines that start parallel to the red
line and move inwards until tangent to the polygon.

I also added some text to explicitly say that the green and blue
highlighted edges are left and right bounding edges.